🔨 Make the dwin driver host-buildable, and fix the pattern that blocked it
The third LCD driver now compiles and runs under the test HAL — config 010-dwin, 650
tests green. It had been filed as "not host-buildable" since early on this branch,
alongside two others, and the three turn out to have nothing in common.

Three blockers, each smaller than the label suggested.

LCD_SERIAL: shared/serial_ports.h wants MSERIAL(n) and this HAL has no numbered-port
machinery, so it gets a second named MSerialT. That is not only a fix — a DWIN panel
is driven by writing bytes at a screen rather than by calling an interface, so the
byte stream is the only observable there is, and a real second port is what would let
a test assert what the firmware told the display.

WString.h: the driver includes it and never uses String — checked with a
word-boundary grep, every hit is dwinDrawString taking a char*. The shim is therefore
deliberately empty. Providing a real String would let a later edit start depending on
heap allocation in firmware that has none, with no complaint from the compiler.

dwinDrawSignedFloat "ambiguity": not an ambiguity at all. dwin.h declares it taking
long, dwin.cpp defines it taking int32_t, and on arm-none-eabi those are the same
type, so the header has been describing a function that does not exist anywhere else.
Corrected to match the definition, and proven inert on the shipping target by a
controlled before-and-after build in one session: 126412 bytes of flash both ways. An
earlier comparison across sessions differed by 8 bytes and would have supported
whichever conclusion I preferred.

That made three instances of one pattern in a session, so it is now register #38: a
declaration and its definition differing by a typedef and agreeing on exactly one
platform. The third instance was in this fork's own test HAL —
delayMicroseconds(unsigned long) declared against a uint32_t definition, an undefined
reference waiting for the first caller outside the usual configurations. Instrument
defect, fixed on sight.

And the new port immediately reproduced the documented hang: nothing drains a display
port, so the first refresh that fills the 128-byte buffer spins forever and the suite
hangs instead of failing. Marked as having no host attached, in the same translation
unit as the port and after it, because initialisation order is only guaranteed within
a TU and register #20 was exactly that mistake.

All ten configs green: 604 605 612 674 675 613 610 617 670 650.

+++ Marlin/src/HAL/TEST/include/Arduino.h
@@ -340,7 +340,10 @@ extern "C" {
 
 // Time functions
 extern "C" void delay(const int ms);
-void delayMicroseconds(unsigned long);
+// uint32_t, matching the definition in arduino.cpp. Declared `unsigned long` it is a
+// different function on LP64 and the link fails — which nothing noticed until a driver
+// outside the usual set of test configurations called it.
+void delayMicroseconds(uint32_t);
 unsigned long millis();
 
 // IO functions
